update core plugins
also move spawning of main camera to the main camera plugin
