+++ crates/nora_core/src/plugins.rs
@@ -4,13 +4,13 @@ pub(crate) mod main_camera;
 pub mod physics;
 
 use bevy::app::{PluginGroup, PluginGroupBuilder};
-use bevy::math::Vec3;
+use nora_object_interaction::InteractionPlugin;
 use crate::plugins::{
     core::CorePlugin,
     menu::MenuPlugin,
     main_camera::MainCameraPlugin,
 };
-use nora_physics::PhysicsPlugin;
+use crate::plugins::physics::DefaultPhysicsPlugin;
 
 
 pub struct CorePlugins;
@@ -20,6 +20,7 @@ impl PluginGroup for CorePlugins {
         group.add(CorePlugin);
         group.add(MenuPlugin);
         group.add(MainCameraPlugin);
-        group.add(PhysicsPlugin{gravity: Vec3::new(0.0, -9.81, 0.0)});
+        group.add(DefaultPhysicsPlugin);
+        group.add(InteractionPlugin);
     }
 }
